refactor: move deploy functions under src
